requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
confluent-kafka==2.10.0
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
